requirements.txt
asgiref==3.5.2
black==22.10.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
Django==4.1.1
django-choices==1.7.2
django-extensions==3.2.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycparser==2.21
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
